Database is working and prepopulating with categories and questions now, yay. Could use further testing. Maybe there is a more efficient way?

+++ app/src/main/java/com/example/project02_triviaapp/database/entities/Category.java
@@ -9,7 +9,7 @@ import java.util.Objects;
 @Entity(tableName = "category_table")
 public class Category {
     @PrimaryKey(autoGenerate = true)
-    private int categoryId;
+    private long categoryId;
     private String categoryText;
 
     public Category(@NonNull String categoryText) {
@@ -29,11 +29,11 @@ public class Category {
         return Objects.hash(categoryId, categoryText);
     }
 
-    public int getCategoryId() {
+    public long getCategoryId() {
         return categoryId;
     }
 
-    public void setCategoryId(int categoryId) {
+    public void setCategoryId(long categoryId) {
         this.categoryId = categoryId;
     }
